adding mappers

+++ src/main/java/com/solvd/db/model/Card.java
@@ -21,7 +21,7 @@ public class Card {
         return cardNumber;
     }
 
-    public void setCardNumber(int cardNumber) {
+    public void setCardNumber(long cardNumber) {
         this.cardNumber = cardNumber;
     }
